requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
pillow>=10.0.0
accelerate>=0.24.0
compressed-tensors>=0.1.0
google-generativeai>=0.3.0
tokenizers>=0.19.0
huggingface-hub>=0.20.0

